requirements.txt
loguru==0.7.2
shapely==2.0.6
joblib==1.4.2
pydantic==2.9.2
pandas==2.2.3
pytest==8.3.3
fiftyone==1.0.1
# torch==2.5.1
exif==1.6.0
rasterio>=1.4.2,<1.6
pyproj==3.7.0
jupyter==1.1.1
#  pip install fiftyone pandas pydantic joblib shapely loguru pytest
dvc==3.57.0
pillow~=11.2.1
pandas~=2.2.3
geopandas~=1.1.0
#sahi==0.11.20
# osgeo
piq==0.8.0
#umap-learn
contextily
matplotlib_scalebar==0.9.0
adjustText==1.3.0
matplotlib-map-utils==3.0.0
rioxarray==0.19.0
networkx~=3.5
pyarrow~=20.0.0
pycocotools~=2.0.9
opencv-contrib-python~=4.11
rtree~=1.4.0
#kornia~=0.8.1
albumentations==1.0.3 # TODO update this
wandb~=0.20.1 # TODO update this
# ultralytics
timm==1.0.19
scipy==1.16.1
numpy==1.26.4